Updates for DevPreview: Refactor Composite* classes into unit-of-work package
