--- src/api/composite-api/index.ts
@@ -1,11 +1,0 @@
-export * from './CompositeApi';
-export * from './CompositeRequest';
-export * from './CompositeSubrequest';
-
-export enum Method {
-    DELETE = 'DELETE',
-    GET = 'GET',
-    PATCH = 'PATCH',
-    POST = 'POST',
-    PUT = 'PUT',
-}
