+++ src/api/index.ts
@@ -1,4 +1,3 @@
 export * from './ConnectionConfig';
-export * from './composite-api'
 export * from './ForceApi';
 export * from './unit-of-work';
