+++ src/api/unit-of-work/CompositeRequest.ts
@@ -1,6 +1,5 @@
 import * as _ from 'lodash';
-
-import { CompositeSubrequest } from './';
+import {CompositeSubrequest} from './CompositeSubrequest';
 
 export class CompositeRequest {
     private allOrNone: boolean;
