+++ src/api/unit-of-work/index.ts
@@ -1,1 +1,9 @@
 export { UnitOfWork, UnitOfWorkResponse, UnitOfWorkResult } from './UnitOfWork';
+
+export enum Method {
+    DELETE = 'DELETE',
+    GET = 'GET',
+    PATCH = 'PATCH',
+    POST = 'POST',
+    PUT = 'PUT',
+};
