+++ src/functions/index.ts
@@ -23,7 +23,6 @@ export class Context {
         public readonly apiVersion: string,
         public readonly userContext: UserContext,
         public readonly logger: Logger,
-        public readonly payloadVersion: string,
         public readonly forceApi?: ForceApi,
         public readonly unitOfWork?: UnitOfWork
         ) { }
